Begin split of supportFunctions.py to specialized modules.
